backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
python-jose==3.3.0
passlib==1.7.4
python-dotenv==1.0.1
pydantic-settings==2.2.1
httpx==0.27.0
pytest==8.2.0
chromadb==0.5.23
sentence-transformers==5.2.3
groq==0.37.1
python-multipart==0.0.9
bcrypt==4.0.1
redis==7.3.0
langchain-core==1.2.19
langchain-groq==1.1.2
langgraph==1.1.2
langgraph-checkpoint==4.0.1
langgraph-prebuilt==1.0.8
langgraph-sdk==0.3.11
geoalchemy2==0.18.4
shapely==2.0.4
rq==2.7.0
rq-scheduler==0.14.0